Downloads/senseticket/requirements.txt
# requirements.txt
# Dependencies for SENSE Bot v2

# Discord
discord.py>=2.3.0


# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Web Framework
flask>=3.0.0

# AI & Analysis
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
PyNaCl>=1.5.0

# Optional: For embeddings (if needed)
# sentence-transformers>=2.2.0

# Optional: For caching
# redis>=5.0.0